Use winston logger not console logger in prod explorer setup
