+++ explorer/src/server.ts
@@ -1,6 +1,5 @@
 import * as controllers from './controllers'
-import { requestWhitelist, logger, errorLogger } from 'express-winston'
-import * as winston from 'winston'
+import { addRequestLogging, logger } from './logging'
 import express from 'express'
 import http from 'http'
 import mime from 'mime-types'
@@ -9,34 +8,6 @@ import seed from './seed'
 
 export const DEFAULT_PORT = parseInt(process.env.SERVER_PORT, 10) || 8080
 
-const LOGGER_WHITELIST = [
-  'url',
-  'method',
-  'httpVersion',
-  'originalUrl',
-  'query'
-]
-requestWhitelist.splice(0, requestWhitelist.length, ...LOGGER_WHITELIST)
-
-const addLogging = (app: express.Express) => {
-  const consoleTransport = new winston.transports.Console()
-
-  app.use(
-    logger({
-      expressFormat: true,
-      meta: true,
-      msg: 'HTTP {{req.method}} {{req.url}}',
-      transports: [consoleTransport]
-    })
-  )
-
-  app.use(
-    errorLogger({
-      transports: [consoleTransport]
-    })
-  )
-}
-
 const server = (port: number = DEFAULT_PORT) => {
   if (process.env.NODE_ENV === 'development') {
     seed()
@@ -44,7 +15,7 @@ const server = (port: number = DEFAULT_PORT) => {
 
   const app = express()
   if (process.env.NODE_ENV !== 'test') {
-    addLogging(app)
+    addRequestLogging(app)
   }
 
   app.use(
@@ -66,7 +37,7 @@ const server = (port: number = DEFAULT_PORT) => {
   const server = new http.Server(app)
   bootstrapRealtime(server)
   return server.listen(port, () => {
-    console.log(`server started, listening on port ${port}`)
+    logger.info(`server started, listening on port ${port}`)
   })
 }
